Specify ABI for assembler
For some reason you need to specify what ABI you are targeting.
I guess the configure script for the GCC toolchain does not
propagate properly (might be a bug).

Also the assembler rule for the lowercase 's' was removed.
Forgot why I placed it there but I don't think it is needed.