Refactor build system + directory structure
Okay this is a huge commit but and goes against the spirit of
"one change per commit" but I was too lazy to split things up
so I did a premature squash.

1. Changed directory structure
It now more closely resembles most linux source trees. I figured it
would be easier to manage the chaos with a more tried and true layout.

2. Use nested makefiles
Again, borrowing from other kernel projects; modules now have their
own Makefile. There is one Makefile to build the whole kernel image
in the top level directory of the project.

In the near future we should consider adding configuration files
to further improve the flexiblity.

3. Out-of-source building
For some reason this has been challenging to implement, mainly due
to Makefile syntax woes. But I finally got it to work so now the
source tree will remain clean.