Fix compile issue with csr functions
Apparently the riscv assembler is running behind the privileged spec
and is still using the older names for certain regeisters. So when
attempting to compile, we get "csxx requires and absolute expression".
Now I know this just means that the register referenced is non-existent
or is not compatible with the chosen opcode.