Fix memory corruption on blocking ipc calls
As it turns out, blocking threads is somewhat complicated
due to managing the state of the processor and threads.
Since we were using a single kernel stack for all threads and
the kernel stack pointer is reset, a blocked thread would have
its stack frames corrupted by a subseqent entry into the main
trap handler.